Move TestIsWatchlist to testidentification pkg
